requirements.txt
fastapi
uvicorn
motor
pymongo
passlib[bcrypt]
email-validator
bcrypt==4.0.1
pydantic
python-jose[cryptography]
python-multipart
python-dotenv
PyPDF2 
pdfplumber
pytesseract 
pillow 
pdf2image 
opencv-python
easyocr
sentence-transformers 
faiss-cpu
numpy
PyMuPDF
openai 
diffusers 
transformers 
accelerate 
torch 
torchvision 
requests 


